#30: refac. der Apothekenklasse
