+++ src/main/resources/kunden.txt
@@ -5,3 +5,5 @@
 
 68,test,test,test,test@email.com,test
 805,a,a,a,a,a
+694,s,s,s,s,s
+336,f,ff,f,f,f
